консольный обозреватель сообщений

+++ Observer/ConnResetEvent.cpp
@@ -1,6 +1,6 @@
 #include "ConnResetEvent.h"
 
-ConnResetEvent::ConnResetEvent(int con_num)
+ConnResetEvent::ConnResetEvent(size_t con_num)
     : _con_num(con_num)
 {}
 
@@ -9,7 +9,7 @@ eventType ConnResetEvent::GetEventType() const
     return connReset;
 }
 
-int ConnResetEvent::GetNumber() const
+size_t ConnResetEvent::GetNumber() const
 {
     return _con_num;
 }
